Add mutation logic into its own component
